Add territory effect movement cost (#5234)
* Add new territory effect option and helper methods

* Update route finding to use movement costs

* Update route finding and creation code to use route by movement cost

* Update delegates and util to use new route methods

* Delete unused composite route finder

* Update UI code to handle new route methods

* Update Pro AI to use new route methods

* Update weak AI for some refactoring

* Update unit tests to handle new route methods

* Remove unnecessary unit test

* Fix format violations

* GameMap refactoring

* Fix some BigDecimal checks

* Remove unneeded parenthesis

* Add property for using remaining movement and a few fixes

* Fix already move change

* Fix condition

* Add getMaxMovementCost unit tests

* Add additional getMaxMovementCost unit test

* Update fuel cost to consider movement cost

* Fix java doc typo

* Refactoring updates

* Remove unnecessary method

* Fix formatting

* Refactor unit tests

+++ game-core/src/main/java/games/strategy/triplea/ai/pro/ProCombatMoveAi.java
@@ -1397,7 +1397,7 @@ public class ProCombatMoveAi {
               Matches.territoryHasNeighborMatching(
                       data, Matches.territoryIsInList(ProUtils.getLiveAlliedCapitals(data, player)))
                   .test(t);
-          final int range = TripleAUnit.get(unit).getMovementLeft();
+          final int range = TripleAUnit.get(unit).getMovementLeft().intValue();
           final int distance =
               data.getMap()
                   .getDistance_IgnoreEndForCondition(
@@ -1470,7 +1470,7 @@ public class ProCombatMoveAi {
               Matches.territoryHasNeighborMatching(
                       data, ProMatches.territoryHasInfraFactoryAndIsAlliedLand(player, data))
                   .test(t);
-          final int range = TripleAUnit.get(unit).getMovementLeft();
+          final int range = TripleAUnit.get(unit).getMovementLeft().intValue();
           final int distance =
               data.getMap()
                   .getDistance_IgnoreEndForCondition(
@@ -2047,7 +2047,7 @@ public class ProCombatMoveAi {
         Matches.territoryHasNeighborMatching(
                 data, ProMatches.territoryHasInfraFactoryAndIsAlliedLand(player, data))
             .test(t);
-    final int range = TripleAUnit.get(unit).getMovementLeft();
+    final int range = TripleAUnit.get(unit).getMovementLeft().intValue();
     final int distance =
         data.getMap()
             .getDistance_IgnoreEndForCondition(
